+++ src/pages/authentication/common/hooks.ts
@@ -31,7 +31,7 @@ export interface LoginPrecheck {
 export function useLoginPrecheck() {
   return (email: string) => {
     return request('POST', endpoint('/api/v1/login/precheck'), { email }).then(
-      (response: AxiosResponse<LoginPrecheck>) => response.data
+      (response: AxiosResponse<{ data: LoginPrecheck }>) => response.data.data
     );
   };
 }
